+++ app/resolver/environ.js
@@ -1,5 +1,9 @@
 var EnvironResolver = require('../../lib/environresolver');
 
-exports = module.exports = function() {
-  return new EnvironResolver();
+exports = module.exports = function(registry) {
+  return new EnvironResolver(registry);
 };
+
+exports['@require'] = [
+  'module:bixby-ns.ServiceRegistry'
+];